DB Scaffold på Order databasen.

+++ FroSidanMVC/Models/Entities/Product.cs
@@ -9,8 +9,8 @@ namespace FroSidanMVC.Models.Entities
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Price { get; set; }
-        public int? TempPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal? TempPrice { get; set; }
         public int Balance { get; set; }
         public string ImgRef { get; set; }
         public string ThumbRef { get; set; }
